Fix disconnect Bye-flush race, file-consent HOL-block, stop idempotency
Five bugs from the production-readiness audit, all in the host/
controller session core:

1. ControllerSession::disconnect sent Bye into a channel then
   immediately aborted the control task with no grace period for it to
   actually flush ControlMsg::Bye onto the wire — unlike
   HostSession::stop/emergency_stop, which both grace-wait via
   BYE_FLUSH_GRACE before falling back to abort. Every ordinary
   "Disconnect" click was landing as a hard transport drop, wasting the
   full reconnect window on the host side. Fixed by applying the same
   grace-then-abort pattern.

2. The file-transfer consent prompt ran inline inside
   host_control_loop's select!, head-of-line-blocking a live
   controller's Input/KeyframeRequest/Bye for up to 90s while the local
   user decided on a file offer — exactly the bug class already fixed
   for ControlRequest via the consent_rx pattern, but never applied
   here. Fixed by splitting host_handle_file_offer into a precheck
   (sync) and a finish (post-consent) phase, spawning the consent prompt
   off-loop and delivering its result back via a new file_consent_rx
   channel, mirroring ControlRequest's existing wiring exactly. All
   original side effects (authorization, per-transfer consent,
   FileAccept/Reject, audit, lifecycle events) are unchanged, only
   relocated to the correct phase.

3. The inbound peer-Bye handler used a plain store(true) instead of the
   swap-based first-caller-wins guard stop()/emergency_stop() both use,
   so a session already locally emergency-stopped could still run this
   arm's full teardown and emit a second, contradictory SessionEnded
   event. Fixed to use the same swap-and-check pattern.

4. host_handle_file_complete discarded SafeFileWriter::finish()'s fsync
   error, so a flush failure (e.g. disk-full) let the host report a
   clean transfer completion. Now aborts and sends FileReject on error.

5. abr.rs's bandwidth-utilisation multiply ran in f32, which can lose
   precision above ~16.7M silently. Fixed by promoting the constant and
   the multiply to f64 (not just casting after — an f32 constant bakes
   in its own rounding).

Fixing #1 made a previously-racy test ordering deterministic, which
surfaced a genuine pre-existing gap: join_audio/abort_cursor/
abort_file_transfer weren't guaranteed to run on every control-loop exit
path. Extended the existing once-only teardown wrapper (already used for
input-flush) to cover those too, so teardown happens exactly once
regardless of which path gets there first.

Found by a 5-agent production-readiness audit, fixed by a dedicated
agent and independently re-verified — including a standalone regression
test proving #2's fix empirically (a chat message surfaces in ~10-20ms
during a pending file-consent prompt, rather than being blocked behind
it). 76/76 tests passing, repeated 4x with no flakiness; cargo check
--workspace and the app crate both clean (no broken downstream caller
from the private-function signature changes).

+++ crates/ras-core/src/abr.rs
@@ -15,8 +15,10 @@ use ras_transport_iroh::ConnHealth;
 const LOSS_BACKOFF_THRESHOLD: f32 = 0.02;
 /// Multiplicative decrease applied when loss exceeds the threshold (AIMD-style).
 const BACKOFF_FACTOR: f32 = 0.8;
-/// Fraction of the estimated deliverable rate we allow ourselves to fill (headroom for pacing).
-const BANDWIDTH_UTILISATION: f32 = 0.9;
+/// Fraction of the estimated deliverable rate we allow ourselves to fill (headroom for pacing). `f64`
+/// (not `f32`) so the multiply below stays exact — an `f32` constant would bake in its own rounding
+/// (e.g. `0.9f32` is actually `0.899999976...`) before the deliverable-bandwidth multiply even runs.
+const BANDWIDTH_UTILISATION: f64 = 0.9;
 
 /// SVC layer-shedding thresholds, as a fraction of the SESSION ceiling (`self.ceiling_bps`, not the
 /// per-tick bandwidth-constrained one) that the current bandwidth-derived tick ceiling has fallen to.
@@ -76,8 +78,11 @@ impl AdaptiveBitrateController for LatencyFirstAbr {
         feedback: Option<DecoderFeedback>,
     ) -> BitrateDecision {
         // Ceiling for *this* tick: the smaller of the session ceiling and a fraction of what the
-        // path can actually deliver. `as` truncation is fine — bitrates are far below u32::MAX here.
-        let deliverable = (health.estimated_bandwidth_bps as f32 * BANDWIDTH_UTILISATION) as u32;
+        // path can actually deliver. The multiply runs in f64 (not f32) so values above f32's 24-bit
+        // mantissa (~16.7 Mbps) don't silently lose precision before the final truncating cast back to
+        // u32 — bitrates here are far below u32::MAX, so the cast itself is fine.
+        let deliverable =
+            (f64::from(health.estimated_bandwidth_bps) * BANDWIDTH_UTILISATION) as u32;
         let tick_ceiling = self.ceiling_bps.min(deliverable).max(self.floor_bps);
 
         if health.loss_fraction > LOSS_BACKOFF_THRESHOLD {
